Feat: added category module
