+++ resources/views/categories/action.blade.php
@@ -1,2 +1,2 @@
-<a href="javascript:void(0)" class="edit btn btn-primary btn-sm">Edit</a>
+<a href="javascript:void(0)" class="edit btn btn-primary btn-sm category-edit-btn" data-id="{{$id}}">Edit</a>
 <a href="javascript:void(0)" class="edit btn btn-danger btn-sm" id="categoryDeleteBtn" data-id="{{$id}}">Delete</a>
